ui/sdl2: release grab even when window is scaled
In absolute mode, we want to release the mouse grab when the user moves
the mouse outside of the window. Previously, if the SDL2 ui was being
used with OpenGL disabled, and the window was larger than the size
of the guest surface, we would not release the grab when the mouse
is at the bottom or right edge of the window.

This is because in this case we use SDL_RenderSetLogicalSize, which
scales the coordinate system used by the mouse events.

Signed-off-by: Maja Kądziołka <maya@compilercrim.es>